Allow pypng to be unbundled.
